# static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Wrembly Image API",
    "description": "API for image processing and OCR",
    "version": "1.0.0"
  },
  "basePath": "/",
  "schemes": [
    "https"
  ],
  "paths": {
    "/": {
      "get": {
        "summary": "Welcome endpoint",
        "responses": {
          "200": {
            "description": "Welcome message"
          }
        }
      }
    },
    "/api/upload": {
      "post": {
        "summary": "Upload an image to Azure Storage",
        "parameters": [
          {
            "in": "formData",
            "name": "file",
            "type": "file",
            "required": true,
            "description": "The image file to upload"
          }
        ],
        "responses": {
          "200": {
            "description": "File uploaded successfully"
          },
          "400": {
            "description": "No file part or no selected file"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/images": {
      "get": {
        "summary": "List all images in Azure Storage",
        "responses": {
          "200": {
            "description": "List of images"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/images/{filename}": {
      "delete": {
        "summary": "Delete an image from Azure Storage",
        "parameters": [
          {
            "in": "path",
            "name": "filename",
            "type": "string",
            "required": true,
            "description": "The filename to delete"
          }
        ],
        "responses": {
          "200": {
            "description": "File deleted successfully"
          },
          "404": {
            "description": "File not found"
          }
        }
      }
    },
    "/api/roi-info": {
      "post": {
        "summary": "Update ROI information file",
        "parameters": [
          {
            "in": "formData",
            "name": "file",
            "type": "file",
            "required": true,
            "description": "The ROI info file to upload"
          }
        ],
        "responses": {
          "200": {
            "description": "ROI info updated successfully"
          },
          "400": {
            "description": "Invalid file or no file part"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "get": {
        "summary": "Get ROI information",
        "responses": {
          "200": {
            "description": "ROI information content"
          },
          "404": {
            "description": "ROI info file not found"
          }
        }
      }
    },
    "/api/ocr": {
      "post": {
        "summary": "Process OCR on an image using ROI coordinates based on current set order",
        "parameters": [
          {
            "in": "formData",
            "name": "file",
            "type": "file",
            "required": true,
            "description": "The image file to process OCR"
          }
        ],
        "responses": {
          "200": {
            "description": "OCR results",
            "schema": {
              "type": "object",
              "properties": {
                "set_order": {
                  "type": "integer",
                  "description": "Current set order used for processing"
                },
                "results": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "roi_index": {
                        "type": "integer",
                        "description": "Index of the ROI"
                      },
                      "coordinates": {
                        "type": "array",
                        "items": {
                          "type": "integer"
                        },
                        "description": "ROI coordinates [x1, y1, x2, y2]"
                      },
                      "text": {
                        "type": "string",
                        "description": "OCR result text"
                      }
                    }
                  }
                },
                "result_file": {
                  "type": "string",
                  "description": "Name of the saved result file"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/ocr-results/{filename}": {
      "get": {
        "summary": "Get OCR result from a specific file",
        "parameters": [
          {
            "in": "path",
            "name": "filename",
            "type": "string",
            "required": true,
            "description": "The OCR result file to retrieve"
          }
        ],
        "responses": {
          "200": {
            "description": "OCR result content"
          },
          "404": {
            "description": "Result file not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/set-order": {
      "post": {
        "summary": "Update set order value",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "type": "object",
              "required": ["value"],
              "properties": {
                "value": {
                  "type": "integer",
                  "description": "The new set order value"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Set order updated successfully"
          },
          "400": {
            "description": "Invalid input"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "get": {
        "summary": "Get current set order value",
        "responses": {
          "200": {
            "description": "Current set order value"
          },
          "404": {
            "description": "Set order not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    }
  }
} 